+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 <a name="0.2.0"></a>
 
+- [table with searchBar](https://github.com/Capgemini/dcx-react-library/issues/193)
 - [table](https://github.com/Capgemini/dcx-react-library/issues/189)
 - [tabGroup](https://github.com/Capgemini/dcx-react-library/issues/167)
 - [copyToClipboard](https://github.com/Capgemini/dcx-react-library/issues/184)